:fire: Remove unwanted utils
